FEAT: Loaded the repo with the current code base
GameLogic houses all the basic logic of the Pinochle games.
This includes cards, decks, players, scoring, playing
hands, etc. (all for partners Pinochle).

NeuralNetModels houses the code for creating neural nets
that will learn to play Pinochle. This includes code for
building datasets, and it includes the datasets themselves.

DesktopApp houses the desktop version of Pynochle,
written in PyQt5. This includes a bunch of stolen
playing card icons.

WebApp will house the code for a browser-based version
of Pynochle.