FIX: Changed hash calculation for Card back to old style
Now, the hash of two cards are equal if the cards are equal.
Thus, we can have several different, distinguishable
instances of the jack of spades, and yet their hashes are
all equal.

This allows us to use the jack of diamonds to index into
a dictionary, and any jack of diamonds will get us access
to the same dict value.

The card_players dictionary usage had to be refactored to
allow for this hash method.